fix ugc perm

+++ app/Http/Controllers/Admin/BlogController.php
@@ -34,6 +34,7 @@ class BlogController extends Controller
 
     public function edit(Blog $blog)
     {
+        $this->authorize('update', $blog);
         return view("cms.blog.v_create_blog", compact("blog"));
     }
 
@@ -49,6 +50,8 @@ class BlogController extends Controller
 
     public function update(BlogRequest $request)
     {
+        $blog = Blog::findOrFail($request->id);
+        $this->authorize('update', $blog);
         $createBlog = $this->blogService->update($request->validated());
         if ($createBlog) {
             return $this->success($createBlog, 'Berita berhasil dibuat');
@@ -59,6 +62,8 @@ class BlogController extends Controller
 
     public function destroy($id)
     {
+        $blog = Blog::findOrFail($id);
+        $this->authorize('delete', $blog);
         $data = $this->blogService->delete($id);
         return $this->success($data, 'Data berhasil dihapus');
     }
